ai-service/requirements.txt
flask
flask-cors
# psycopg2-binary removed for sqlite
redis
scikit-learn
pandas
numpy
openai
python-dotenv
requests
